refactor: make identity edit modal schema-aware and eliminate code duplication
- Extract shared form widgets (TelWidget, TextWidget, templates) to separate file
- Replace hardcoded fields with dynamic schema-based form generation in edit modal
- Fix incorrect field mapping that prevented editing identities with different schemas
